requirements.txt
Pillow==6.0.0
noise==1.2.2
numpy==1.18.4